safari-600.7-branch/LayoutTests/fast/dom/shadow/pseudoclass-update-disabled-option-expected.txt

When an option element became 'disabled' or not-'disabled', distribution should happen.

Since an option element does not create a renderer, we cannot check this using reftest.

PASS nodes1.length is 1
PASS nodes1.item(0).innerHTML is "option 2"
PASS nodes2.length is 1
PASS nodes2.item(0).innerHTML is "option 3"
PASS successfullyParsed is true

TEST COMPLETE
